+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 ## Version 0.36.1-SNAPSHOT
 
 _TBA_
-
-* Fix: Fix showing vertical scrollbar with `VerticalScrollSpec`.
+* Fix: Fix showing vertical scrollbar with `VerticalScrollSpec`. Default behaviour is scrollbars disabled.
 
 For more details, see the [full diff](https://github.com/facebook/litho/compare/v0.36.0...master).