+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 _TBA_
 * New: Add `.duplicateChildrenStates(boolean)` method to `Component` which passes the flag to [`ViewGroup#setAddStatesFromChildren(boolean)`](https://developer.android.com/reference/android/view/ViewGroup#setAddStatesFromChildren(boolean)). When this flag is set to true, the component applies all of its children's drawable states (focused, pressed, etc.) to itself.
+* New: Ability to specify the percentage of Component's width/height which should be visible to trigger Visible events. Read more in [the documentation](https://fblitho.com/docs/visibility-handling#custom-visibility-percentage).
 * Fix: Fix showing vertical scrollbar with `VerticalScrollSpec`. Default behaviour is scrollbars disabled.
 
 For more details, see the [full diff](https://github.com/facebook/litho/compare/v0.36.0...master).