+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Version 0.37.1-SNAPSHOT
+
+_release-date_
+
+* TBA
+
+For more details, see the [full diff](https://github.com/facebook/litho/compare/v0.37.0...master).
+
+
 ## Version 0.37.0
 
 _2020-07-07_